backend/core-platform/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.1
firebase-admin==6.1.0
requests==2.30.0
